moved json/yaml unmarshaling into their own files
This will give us more flexibility to ensure unmarshaling in both yaml/json
returns a consistent structure.
