+++ config/config_test.go
@@ -291,11 +291,11 @@ var v03OpenTelemetryConfig = OpenTelemetryConfiguration{
 			{Name: "service.name", Value: "unknown_service"},
 			{Name: "string_key", Type: &AttributeNameValueType{Value: "string"}, Value: "value"},
 			{Name: "bool_key", Type: &AttributeNameValueType{Value: "bool"}, Value: true},
-			{Name: "int_key", Type: &AttributeNameValueType{Value: "int"}, Value: float64(1)},
+			{Name: "int_key", Type: &AttributeNameValueType{Value: "int"}, Value: 1},
 			{Name: "double_key", Type: &AttributeNameValueType{Value: "double"}, Value: 1.1},
 			{Name: "string_array_key", Type: &AttributeNameValueType{Value: "string_array"}, Value: []interface{}{"value1", "value2"}},
 			{Name: "bool_array_key", Type: &AttributeNameValueType{Value: "bool_array"}, Value: []interface{}{true, false}},
-			{Name: "int_array_key", Type: &AttributeNameValueType{Value: "int_array"}, Value: []interface{}{float64(1), float64(2)}},
+			{Name: "int_array_key", Type: &AttributeNameValueType{Value: "int_array"}, Value: []interface{}{1, 2}},
 			{Name: "double_array_key", Type: &AttributeNameValueType{Value: "double_array"}, Value: []interface{}{1.1, 2.2}},
 		},
 		AttributesList: ptr("service.namespace=my-namespace,service.version=1.0.0"),
